Add an implicit capability for each component
This commit makes sure that every component has an implicit capability corresponding
to its GAV coordinates. For example, the module `asm:asm:3.0` has an implicit capability
with the same group, name and version. This simplifies rules, in particular when a module
needs to explain that it provides the same capability as a different module published
at different coordinates.